requirements.txt
asgiref==3.5.0
boto3==1.23.10
botocore==1.26.10
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-crispy-forms==1.14.0
django-excel==0.0.10
django-import-export==2.8.0
django-storages==1.12.3
engineering-notation==0.6.0
et-xmlfile==1.1.0
idna==3.3
jmespath==1.0.0
lml==0.1.0
MarkupPy==1.14
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.9
panda==0.3.1
pandas==1.4.2
Pillow==9.1.0
psycopg2==2.9.3
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-ods==0.6.0
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pyexcel-xlsxw==0.6.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
texttable==1.6.4
tk-tools==0.15.0
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
